requirements.txt
mkdocs>=1.5.0
mkdocs-material>=9.4.0
mkdocs-jupyter>=0.24.0
mkdocs-with-pdf>=0.9.3
pytest>=7.4.0
hypothesis>=6.82.0
pymdown-extensions>=10.2
mkdocs-material-extensions>=1.2
python-frontmatter>=1.0.0
